+++ src/test/scala/io/ergodex/core/Helpers.scala
@@ -1,10 +1,9 @@
 package io.ergodex.core
 
-import io.ergodex.core.syntax.Coll
+import io.ergodex.core.syntax.{Coll, CollOpaque}
 
 object Helpers {
+  def tokenId(s: String): Coll[Byte] = CollOpaque(s.getBytes().toVector)
 
-  def tokenId(s: String): Coll[Byte] = s.getBytes().toVector
-
-  def boxId(s: String): Coll[Byte] = s.getBytes().toVector
+  def boxId(s: String): Coll[Byte] = CollOpaque(s.getBytes().toVector)
 }
