+++ src/test/scala/io/ergodex/core/RuntimeState.scala
@@ -1,4 +1,5 @@
 package io.ergodex.core
+
 import cats.Functor
 import cats.mtl.Ask
 import cats.mtl.syntax.ask._
@@ -8,12 +9,12 @@ import tofu.syntax.monadic._
 final case class SigmaProp(value: String)
 
 final case class RuntimeCtx(
-  height: Int,
-  vars: Map[Int, Any]                        = Map.empty,
-  inputs: List[Box[BoxRuntime.NonRunnable]]  = List.empty,
-  outputs: List[Box[BoxRuntime.NonRunnable]] = List.empty,
-  signatories: List[SigmaProp]               = List.empty
-)
+                             height: Int,
+                             vars: Map[Int, Any]                        = Map.empty,
+                             inputs: List[BoxSim[BoxRuntime.NonRunnable]]  = List.empty,
+                             outputs: List[BoxSim[BoxRuntime.NonRunnable]] = List.empty,
+                             signatories: List[SigmaProp]               = List.empty
+                           )
 
 object RuntimeCtx extends WithContext.Companion[RuntimeCtx] {
   def init: RuntimeCtx            = at(1)
