--- src/test/scala/io/ergodex/core/sim/Helpers.scala
@@ -1,9 +1,0 @@
-package io.ergodex.core.sim
-
-import io.ergodex.core.syntax.{Coll, CollOpaque}
-
-object Helpers {
-  def tokenId(s: String): Coll[Byte] = CollOpaque(s.getBytes().toVector)
-
-  def boxId(s: String): Coll[Byte] = CollOpaque(s.getBytes().toVector)
-}
